# swagger.yaml
openapi: 3.0.1
info:
  title: API de Pagos
  description: API para gestionar los pagos
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor local de desarrollo
paths:
  /api/payments:
    get:
      summary: Obtener todos los pagos
      description: Devuelve una lista con todos los pagos registrados.
      responses:
        '200':
          description: Lista de pagos obtenida exitosamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentResponseDTO'
        '500':
          description: Error interno en el servidor.
    post:
      summary: Registrar un nuevo pago
      description: Crea y guarda un nuevo pago en el sistema.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequestDTO'
      responses:
        '200':
          description: Pago registrado exitosamente.
        '400':
          description: Solicitud inválida (datos faltantes o incorrectos).
        '500':
          description: Error interno en el servidor.

components:
  schemas:
    PaymentResponseDTO:
      type: object
      properties:
        id:
          type: integer
          example: 1
        cardNumber:
          type: string
          example: "1234567812345678"
        amount:
          type: number
          format: float
          example: 150.5
        paymentDate:
          type: string
          format: date
          example: "2023-10-20"
        description:
          type: string
          example: "Pago mensual tarjeta de crédito"
    PaymentRequestDTO:
      type: object
      properties:
        cardNumber:
          type: string
          example: "1234567812345678"
        amount:
          type: number
          format: float
          example: 150.5
        paymentDate:
          type: string
          format: date
          example: "2023-10-20"
        description:
          type: string
          example: "Pago mensual tarjeta de crédito"